feat(medallion): complete silver 12 models, gold SCD2, warehouse and OLTP fixes plus changelog 0.2.0
- silver: 12 incremental models with dedup, full schema.yml 112 tests, shared incremental_utils macro and generic tests
- gold: SCD2 dims via snapshots, Type1 dims from silver refs, expanded dim_date and fact models, schema rewrites
- bronze: _loaded_date generated columns, reconciled source columns, watermark MERGE fix
- oltp: corrected int/BIGINT types for unit/trailer/distance/weight and added drivers index
- mart: SCD2 driver/vehicle with sk is_current, expanded dims and facts to mirror gold
- orchestration: dbt snapshot before build in DAG and bash/ps scripts
- scripts: env-aware seed URL, powershell strict mode and uv fallback parity, databricks_init fallback

+++ sql/oltp_schema/06_routes.sql
@@ -4,10 +4,10 @@ CREATE TABLE IF NOT EXISTS routes (
     origin_state VARCHAR(10),
     destination_city VARCHAR(100),
     destination_state VARCHAR(10),
-    typical_distance_miles DOUBLE PRECISION,
+    typical_distance_miles INT,
     base_rate_per_mile DOUBLE PRECISION,
     fuel_surcharge_rate DOUBLE PRECISION,
-    typical_transit_days DOUBLE PRECISION,
+    typical_transit_days INT,
     updated_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 CREATE INDEX IF NOT EXISTS idx_routes_updated_at ON routes (updated_at);
